+++ mods/Mining/mod/prototypes/item/stone_spirits.lua
@@ -1,0 +1,131 @@
+require("__RsUtility__.scripts.prototypes")
+
+local bronze_stone_spirit = {
+  name = "rs-bronze-stone-spirit",
+  stack_size = 200,
+  subgroup = "rs-stone-spirit",
+  order = "01",
+  icon = "__RsMining__/graphics/icons/copper_stone_spirit.png",
+  icon_size = 32,
+  icon_mipmaps = 2,
+  pictures = {
+    {
+      filename = "__RsMining__/graphics/icons/copper_stone_spirit.png",
+      size = 32,
+      mipmap_count=2,
+      scale = 0.5
+    },
+    {
+      filename = "__RsMining__/graphics/icons/tin_stone_spirit.png",
+      size = 32,
+      mipmap_count=2,
+      scale = 0.5
+    }
+  }
+}
+create_item(bronze_stone_spirit)
+
+local iron_stone_spirit = {
+  name = "rs-iron-stone-spirit",
+  stack_size = 200,
+  subgroup = "rs-stone-spirit",
+  order = 10,
+  icon = "__RsMining__/graphics/icons/iron_stone_spirit.png",
+  icon_size = 32,
+}
+create_item(iron_stone_spirit)
+
+local steel_stone_spirit = {
+  name = "rs-steel-stone-spirit",
+  stack_size = 200,
+  subgroup = "rs-stone-spirit",
+  order = 20,
+  icon = "__RsMining__/graphics/icons/coal_stone_spirit.png",
+  icon_size = 32,
+}
+create_item(steel_stone_spirit)
+
+local mithril_stone_spirit = {
+  name = "rs-mithril-stone-spirit",
+  stack_size = 200,
+  subgroup = "rs-stone-spirit",
+  order = 30,
+  icon = "__RsMining__/graphics/icons/mithril_stone_spirit.png",
+  icon_size = 32,
+}
+create_item(mithril_stone_spirit)
+
+local adamantite_stone_spirit = {
+  name = "rs-adamantite-stone-spirit",
+  stack_size = 200,
+  subgroup = "rs-stone-spirit",
+  order = 40,
+  icon = "__RsMining__/graphics/icons/adamantite_stone_spirit.png",
+  icon_size = 32,
+}
+create_item(adamantite_stone_spirit)
+
+local runite_stone_spirit = {
+  name = "rs-runite-stone-spirit",
+  stack_size = 200,
+  subgroup = "rs-stone-spirit",
+  order = 50,
+  icon = "__RsMining__/graphics/icons/runite_stone_spirit.png",
+  icon_size = 32,
+}
+create_item(runite_stone_spirit)
+
+local orichalcite_stone_spirit = {
+  name = "rs-orichalcite-stone-spirit",
+  stack_size = 200,
+  subgroup = "rs-stone-spirit",
+  order = 60,
+  icon = "__RsMining__/graphics/icons/orichalcite_stone_spirit.png",
+  icon_size = 32,
+}
+create_item(orichalcite_stone_spirit)
+
+local necrite_stone_spirit = {
+  name = "rs-necrite-stone-spirit",
+  stack_size = 200,
+  subgroup = "rs-stone-spirit",
+  order = 70,
+  icon = "__RsMining__/graphics/icons/necrite_stone_spirit.png",
+  icon_size = 32,
+}
+create_item(necrite_stone_spirit)
+
+local banite_stone_spirit = {
+  name = "rs-banite-stone-spirit",
+  stack_size = 200,
+  subgroup = "rs-stone-spirit",
+  order = 80,
+  icon = "__RsMining__/graphics/icons/banite_stone_spirit.png",
+  icon_size = 32,
+}
+create_item(banite_stone_spirit)
+
+local elder_rune_stone_spirit = {
+  name = "rs-elder-rune-stone-spirit",
+  stack_size = 200,
+  subgroup = "rs-stone-spirit",
+  order = 90,
+  icon = "__RsMining__/graphics/icons/light_animica_stone_spirit.png",
+  icon_size = 32,
+  icon_mipmaps = 2,
+  pictures = {
+    {
+      filename = "__RsMining__/graphics/icons/light_animica_stone_spirit.png",
+      size = 32,
+      mipmap_count=2,
+      scale = 0.5
+    },
+    {
+      filename = "__RsMining__/graphics/icons/dark_animica_stone_spirit.png",
+      size = 32,
+      mipmap_count=2,
+      scale = 0.5
+    }
+  }
+}
+create_item(elder_rune_stone_spirit)