+++ mods/Mining/mod/prototypes/technology/processing_technology.lua
@@ -3,7 +3,6 @@ require("__RsUtility__.scripts.prototypes")
 local iron_bar_tech = {
   name = "rs-iron-processing",
   recipes_to_unlock = {
-    "rs-iron-bar",
     "electronic-circuit",
     "firearm-magazine",
     "inserter",
@@ -26,17 +25,33 @@ local iron_bar_tech = {
   },
   unit = {
     count = 20,
-    time = 5,
-    ingredients = {
-      {name = "mining-1", amount =  1},
-      {name = "smithing-1", amount = 1}
-    }
+    time = 5
   },
   levels = {
     ["mining"] = 10,
     ["smithing"] = 10
   },
+  extra_prerequisites = {"rs-iron-bar"},
   icon = "__RsMining__/graphics/technology/iron-technology.png",
   icon_size = 685
 }
-create_technology(iron_bar_tech)
+create_technology(iron_bar_tech, true)
+
+local steel_bar_tech = {
+  name = "rs-steel-processing",
+  recipes_to_unlock = {
+    "steel-chest"
+  },
+  unit = {
+    count = 50,
+    time = 5
+  },
+  levels = {
+    ["mining"] = 20,
+    ["smithing"] = 20
+  },
+  icon = "__RsMining__/graphics/technology/steel-technology.png",
+  icon_size = 685,
+  extra_prerequisites = {"rs-iron-processing", "rs-steel-bar"}
+}
+create_technology(steel_bar_tech, true)