+++ mods/Utility/mod/data/science_packs.lua
@@ -1,13 +1,13 @@
 return {
-  ["mining-1"] = "rs-bronze-pickaxe",
-  ["mining-2"] = "rs-iron-pickaxe",
-  ["mining-3"] = "rs-steel-pickaxe",
-  ["mining-4"] = "rs-mithril-pickaxe",
-  ["mining-5"] = "rs-adamant-pickaxe",
-  ["mining-6"] = "rs-rune-pickaxe",
-  ["mining-7"] = "rs-orikalkum-pickaxe",
-  ["mining-8"] = "rs-necronium-pickaxe",
-  ["mining-9"] = "rs-bane-pickaxe",
-  ["mining-10"] = "rs-elder-rune-pickaxe"
+  ["mining-1"] = "rs-bronze-stone-spirit",
+  ["mining-10"] = "rs-iron-stone-spirit",
+  ["mining-20"] = "rs-steel-stone-spirit",
+  ["mining-30"] = "rs-mithril-stone-spirit",
+  ["mining-40"] = "rs-adamantite-stone-spirit",
+  ["mining-50"] = "rs-runite-stone-spirit",
+  ["mining-60"] = "rs-orichalcite-stone-spirit",
+  ["mining-70"] = "rs-necrite-stone-spirit",
+  ["mining-80"] = "rs-banite-stone-spirit",
+  ["mining-90"] = "rs-elder-rune-stone-spirit"
 
 }